refactor(cli): extract assistant and live routes
